+++ src/main/java/com/lajos/kossuth/groceries/manager/repository/GroceryListRepository.java
@@ -3,5 +3,5 @@ package com.lajos.kossuth.groceries.manager.repository;
 import com.lajos.kossuth.groceries.manager.model.GroceryList;
 import org.springframework.data.repository.CrudRepository;
 
-public interface GroceryListRepository extends CrudRepository<GroceryList, Integer> {
+public interface GroceryListRepository extends CrudRepository<GroceryList, Long> {
 }
